+++ frontend/src/scenes/dashboard/DashboardTemplateChooser.scss
